+++ hedera-node/hedera-token-service-impl/src/main/java/com/hedera/node/app/service/token/impl/handlers/TokenHandlers.java
@@ -192,7 +192,8 @@ public class TokenHandlers {
                 Objects.requireNonNull(tokenGetNftInfosHandler, "tokenGetNftInfosHandler must not be null");
         this.tokenUpdateNftsHandler =
                 Objects.requireNonNull(tokenUpdateNftsHandler, "tokenUpdateNftsHandler must not be null");
-        this.tokenAirdropsHandler = Objects.requireNonNull(tokenAirdropsHandler, "tokenAirdropsHandler must not be null");
+        this.tokenAirdropsHandler =
+                Objects.requireNonNull(tokenAirdropsHandler, "tokenAirdropsHandler must not be null");
     }
 
     /**
