+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/transactions/crypto/HapiCryptoTransfer.java
@@ -31,7 +31,6 @@ import static java.util.stream.Collectors.toList;
 
 import com.google.common.base.MoreObjects;
 import com.google.protobuf.ByteString;
-import com.google.protobuf.UInt32Value;
 import com.hedera.node.app.hapi.fees.usage.BaseTransactionMeta;
 import com.hedera.node.app.hapi.fees.usage.crypto.CryptoTransferMeta;
 import com.hedera.node.app.hapi.fees.usage.state.UsageAccumulator;
@@ -41,7 +40,7 @@ import com.hedera.node.app.hapi.utils.fee.SigValueObj;
 import com.hedera.node.app.service.evm.utils.EthSigsUtils;
 import com.hedera.services.bdd.spec.HapiSpec;
 import com.hedera.services.bdd.spec.fees.AdapterUtils;
-import com.hedera.services.bdd.spec.transactions.HapiTxnOp;
+import com.hedera.services.bdd.spec.transactions.HapiBaseTransfer;
 import com.hedera.services.bdd.spec.transactions.TxnUtils;
 import com.hedera.services.bdd.spec.transactions.token.TokenMovement;
 import com.hederahashgraph.api.proto.java.AccountAmount;
@@ -50,23 +49,16 @@ import com.hederahashgraph.api.proto.java.CryptoTransferTransactionBody;
 import com.hederahashgraph.api.proto.java.FeeData;
 import com.hederahashgraph.api.proto.java.HederaFunctionality;
 import com.hederahashgraph.api.proto.java.Key;
-import com.hederahashgraph.api.proto.java.NftTransfer;
-import com.hederahashgraph.api.proto.java.TokenID;
 import com.hederahashgraph.api.proto.java.TokenTransferList;
 import com.hederahashgraph.api.proto.java.Transaction;
 import com.hederahashgraph.api.proto.java.TransactionBody;
 import com.hederahashgraph.api.proto.java.TransferList;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicReference;
@@ -74,53 +66,26 @@ import java.util.function.BiConsumer;
 import java.util.function.BinaryOperator;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Predicate;
 import java.util.function.ToLongFunction;
 import java.util.stream.Collector;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class HapiCryptoTransfer extends HapiTxnOp<HapiCryptoTransfer> {
+public class HapiCryptoTransfer extends HapiBaseTransfer<HapiCryptoTransfer> {
     static final Logger log = LogManager.getLogger(HapiCryptoTransfer.class);
 
     private static final List<TokenMovement> MISSING_TOKEN_AWARE_PROVIDERS = Collections.emptyList();
     private static final Function<HapiSpec, TransferList> MISSING_HBAR_ONLY_PROVIDER = null;
 
-    private static final Comparator<AccountID> ACCOUNT_NUM_COMPARATOR = Comparator.comparingLong(
-                    AccountID::getAccountNum)
-            .thenComparingLong(AccountID::getShardNum)
-            .thenComparingLong(AccountID::getRealmNum);
-    private static final Comparator<AccountID> ACCOUNT_NUM_OR_ALIAS_COMPARATOR = (a, b) -> {
-        if (!a.getAlias().isEmpty() || !b.getAlias().isEmpty()) {
-            return ByteString.unsignedLexicographicalComparator().compare(a.getAlias(), b.getAlias());
-        } else {
-            return ACCOUNT_NUM_COMPARATOR.compare(a, b);
-        }
-    };
-    private static final Comparator<AccountAmount> ACCOUNT_AMOUNT_COMPARATOR = Comparator.comparingLong(
-                    AccountAmount::getAmount)
-            .thenComparing(AccountAmount::getAccountID, ACCOUNT_NUM_OR_ALIAS_COMPARATOR);
-    private static final Comparator<NftTransfer> NFT_TRANSFER_COMPARATOR = Comparator.comparing(
-                    NftTransfer::getSenderAccountID, ACCOUNT_NUM_OR_ALIAS_COMPARATOR)
-            .thenComparing(NftTransfer::getReceiverAccountID, ACCOUNT_NUM_OR_ALIAS_COMPARATOR)
-            .thenComparingLong(NftTransfer::getSerialNumber);
-    private static final Comparator<TokenID> TOKEN_ID_COMPARATOR = Comparator.comparingLong(TokenID::getTokenNum);
-    private static final Comparator<TokenTransferList> TOKEN_TRANSFER_LIST_COMPARATOR =
-            (o1, o2) -> Objects.compare(o1.getToken(), o2.getToken(), TOKEN_ID_COMPARATOR);
-
     private boolean logResolvedStatus = false;
     private boolean breakNetZeroTokenChangeInvariant = false;
-
-    private List<TokenMovement> tokenAwareProviders = MISSING_TOKEN_AWARE_PROVIDERS;
     private Function<HapiSpec, TransferList> hbarOnlyProvider = MISSING_HBAR_ONLY_PROVIDER;
     private Optional<String> tokenWithEmptyTransferAmounts = Optional.empty();
     private Optional<Pair<String[], Long>> appendedFromTo = Optional.empty();
     private Optional<AtomicReference<FeeObject>> feesObserver = Optional.empty();
     private Optional<BiConsumer<HapiSpec, CryptoTransferTransactionBody.Builder>> explicitDef = Optional.empty();
-    private boolean fullyAggregateTokenTransfers = true;
     private static boolean transferToKey = false;
 
     @Override
@@ -463,6 +428,7 @@ public class HapiCryptoTransfer extends HapiTxnOp<HapiCryptoTransfer> {
                                 misconfigureIfRequested(b, spec);
                             }
                         });
+
         return builder -> builder.setCryptoTransfer(opBody);
     }
 
@@ -603,141 +569,6 @@ public class HapiCryptoTransfer extends HapiTxnOp<HapiCryptoTransfer> {
             });
             return partyKeys;
         };
-    }
-
-    private List<TokenTransferList> transfersAllFor(final HapiSpec spec) {
-        return Stream.concat(transfersFor(spec).stream(), transfersForNft(spec).stream())
-                .sorted(TOKEN_TRANSFER_LIST_COMPARATOR)
-                .toList();
-    }
-
-    private List<TokenTransferList> transfersFor(final HapiSpec spec) {
-        final Map<TokenID, Pair<Integer, List<AccountAmount>>> aggregated;
-        if (fullyAggregateTokenTransfers) {
-            aggregated = fullyAggregateTokenTransfersList(spec);
-        } else {
-            aggregated = aggregateOnTokenIds(spec);
-        }
-
-        return aggregated.entrySet().stream()
-                .map(entry -> {
-                    final var builder = TokenTransferList.newBuilder()
-                            .setToken(entry.getKey())
-                            .addAllTransfers(entry.getValue().getRight());
-                    if (entry.getValue().getLeft() > 0) {
-                        builder.setExpectedDecimals(
-                                UInt32Value.of(entry.getValue().getLeft().intValue()));
-                    }
-                    return builder.build();
-                })
-                .sorted(TOKEN_TRANSFER_LIST_COMPARATOR)
-                .toList();
-    }
-
-    private Map<TokenID, Pair<Integer, List<AccountAmount>>> aggregateOnTokenIds(final HapiSpec spec) {
-        final Map<TokenID, Pair<Integer, List<AccountAmount>>> map = new HashMap<>();
-        for (final TokenMovement tm : tokenAwareProviders) {
-            if (tm.isFungibleToken()) {
-                final var list = tm.specializedFor(spec);
-
-                if (map.containsKey(list.getToken())) {
-                    final var existingVal = map.get(list.getToken());
-                    final List<AccountAmount> newList = Stream.of(existingVal.getRight(), list.getTransfersList())
-                            .flatMap(Collection::stream)
-                            .sorted(ACCOUNT_AMOUNT_COMPARATOR)
-                            .toList();
-
-                    map.put(list.getToken(), Pair.of(existingVal.getLeft(), newList));
-                } else {
-                    map.put(list.getToken(), Pair.of(list.getExpectedDecimals().getValue(), list.getTransfersList()));
-                }
-            }
-        }
-        return map;
-    }
-
-    private Map<TokenID, Pair<Integer, List<AccountAmount>>> fullyAggregateTokenTransfersList(final HapiSpec spec) {
-        final Map<TokenID, Pair<Integer, List<AccountAmount>>> map = new HashMap<>();
-        for (final TokenMovement xfer : tokenAwareProviders) {
-            if (xfer.isFungibleToken()) {
-                final var list = xfer.specializedFor(spec);
-
-                if (map.containsKey(list.getToken())) {
-                    final var existingVal = map.get(list.getToken());
-                    final List<AccountAmount> newList = Stream.of(existingVal.getRight(), list.getTransfersList())
-                            .flatMap(Collection::stream)
-                            .sorted(ACCOUNT_AMOUNT_COMPARATOR)
-                            .toList();
-
-                    map.put(list.getToken(), Pair.of(existingVal.getLeft(), aggregateTransfers(newList)));
-                } else {
-                    map.put(
-                            list.getToken(),
-                            Pair.of(
-                                    list.getExpectedDecimals().getValue(),
-                                    aggregateTransfers(list.getTransfersList())));
-                }
-            }
-        }
-        return map;
-    }
-
-    private List<AccountAmount> aggregateTransfers(final List<AccountAmount> list) {
-        return list.stream()
-                .collect(groupingBy(
-                        AccountAmount::getAccountID,
-                        groupingBy(AccountAmount::getIsApproval, mapping(AccountAmount::getAmount, toList()))))
-                .entrySet()
-                .stream()
-                .flatMap(entry -> {
-                    final List<AccountAmount> accountAmounts = new ArrayList<>();
-                    for (final var entrySet : entry.getValue().entrySet()) {
-                        final var aa = AccountAmount.newBuilder()
-                                .setAccountID(entry.getKey())
-                                .setIsApproval(entrySet.getKey())
-                                .setAmount(entrySet.getValue().stream()
-                                        .mapToLong(l -> l)
-                                        .sum())
-                                .build();
-                        accountAmounts.add(aa);
-                    }
-                    return accountAmounts.stream();
-                })
-                .sorted(ACCOUNT_AMOUNT_COMPARATOR)
-                .toList();
-    }
-
-    private List<TokenTransferList> transfersForNft(final HapiSpec spec) {
-        final var uniqueCount = tokenAwareProviders.stream()
-                .filter(Predicate.not(TokenMovement::isFungibleToken))
-                .map(TokenMovement::getToken)
-                .distinct()
-                .count();
-        final Map<TokenID, List<NftTransfer>> aggregated = tokenAwareProviders.stream()
-                .filter(Predicate.not(TokenMovement::isFungibleToken))
-                .map(p -> p.specializedForNft(spec))
-                .collect(Collectors.toMap(
-                        TokenTransferList::getToken,
-                        TokenTransferList::getNftTransfersList,
-                        (left, right) -> Stream.of(left, right)
-                                .flatMap(Collection::stream)
-                                .sorted(NFT_TRANSFER_COMPARATOR)
-                                .toList(),
-                        LinkedHashMap::new));
-        if (aggregated.size() != 0 && uniqueCount != aggregated.size()) {
-            throw new RuntimeException("Aggregation seems to have failed (expected "
-                    + uniqueCount
-                    + " distinct unique token types, got "
-                    + aggregated.size()
-                    + ")");
-        }
-        return aggregated.entrySet().stream()
-                .map(entry -> TokenTransferList.newBuilder()
-                        .setToken(entry.getKey())
-                        .addAllNftTransfers(entry.getValue())
-                        .build())
-                .sorted(TOKEN_TRANSFER_LIST_COMPARATOR)
-                .toList();
     }
 
     @Override
