+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/crypto/TokenAirdropSuite.java
@@ -16,12 +16,6 @@
 
 package com.hedera.services.bdd.suites.crypto;
 
-import com.hedera.services.bdd.junit.HapiTest;
-import org.junit.jupiter.api.DynamicTest;
-import org.junit.jupiter.api.Tag;
-
-import java.util.stream.Stream;
-
 import static com.hedera.services.bdd.junit.TestTags.CRYPTO;
 import static com.hedera.services.bdd.spec.HapiSpec.defaultHapiSpec;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
@@ -30,6 +24,11 @@ import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.moving;
 import static com.hedera.services.bdd.suites.HapiSuite.ONE_HUNDRED_HBARS;
 import static com.hederahashgraph.api.proto.java.TokenType.FUNGIBLE_COMMON;
+
+import com.hedera.services.bdd.junit.HapiTest;
+import java.util.stream.Stream;
+import org.junit.jupiter.api.DynamicTest;
+import org.junit.jupiter.api.Tag;
 
 @Tag(CRYPTO)
 public class TokenAirdropSuite {
